requirements.txt
python-dotenv==1.1.1
discord.py==2.6.3
aiohttp==3.12.15
azure-data-tables==12.7.0
